wlan: wait for thread execution to stop before exiting the driver
TX, RX and MC Threads should complete exexution before
the exit routine tries to free the dynamic data these threads
are trying to process. The current implementation does an
interruptible wait for completion of the threads; but it has been
observed that if the process which initiated the driver unload
receives SIGKILL while waiting for threads to complete, the process
will continue without waiting for threads to complete and the threads
will try to access invalidated data after the module has been removed
leading to a kernel panic. To fix this make the wait uninterruptible
and ensure that the driver threads complete execution before the driver
exits.

Change-Id: I3f7e9b4b80b9cc55e6942197bfa30fa41f8e30aa
CRs-Fixed: 550318