wlan: VOS ASSERT in csrRoamCompletion due to Assoc ref count.
While switching between two AP, csr will reissue roam command again
to the nextbss if it was interrupted by the dissconnect req for the
previous bss.During this csr is incrementing bRefAssocStartCnt twice.
When CSR will reissue the roam command it will reset the bRefAssocStartCnt
if it was already set.

CRs-fixed: 554723