wlan: Roam Scan Channels buffer size changed
Currently SIR_ROAM_MAX_CHANNELS is defined as
NUM_RF_CHANNELS. But, NUM_RF_CHANNELS value is
tightly coupled with the nv bin file being used.
To make LFR related scan cache not to be coupled
with other features, the size of SIR_ROAM_MAX_CHANNELS
is defined to a constant value.

Change-Id: Id04d2bf9d75bc56fc2bd261ea593810c2b4e0981
CRs-Fixed: 606397