wlan: prima: Fix ambiguous ioctl error logging
There are multiple places where the message "Invalid IOCTL command" is
logged.  All of the messages are identical, so when the message is
actually logged it is impossible to determine exactly where, and hence
why, it was logged.  Update the logs to add the function name so that
the source of the log will be unique.

Change-Id: I5906627fc3ebd7896dd35e0b82579d02fa915bc3
CRs-fixed: 616348