wlan: Flush/filter temp scan results along with the main results
Flush the temp scan results along with main scan results in
csrScanFlushResult(). This is required to flush the results
completely, as the results are getting updated from temp list to
the main list only after completion of scan.
Filter the temp scan results based on valid channels list along
with the main scan results in csrScanFilterResults().

Change-Id: I54ffe093a4005856d5a96c67869a6b2b710f8c91
CRs-Fixed: 592516