wlan: initialize temp_reg_domain from NV file.
If the temp_reg_domain is not initialzed and the first
regulatory hint comes directly from the CORE when wiphy
is registered then we might end up indexing the wrong
regdomain in the NV structure which leads to memory
corruption. To avoid this initialize temp_reg_domain
with the regulatory domain value obtained from NV
before wiphy initialization.

Change-Id: I6251d08b831af82f69da0ad562b5396bd9e30d8a
CRs-Fixed: 629647