wlan: Disable 40Mhz channels if not supported by Country.
Even if 40Mhz BW is not supported in the Country, the
40Mhz center channels are enabled.
Also If the channel is valid for HT40+ but invalid for HT40- or
vice versa, the 40Mhz center channel is not enabled.

Adds a check to disable 40Mhz center channel if not
supported by Country.
If the channel is valid for either HT40+ or HT40-,
enable the 40Mhz center channel.
Device will follow the BW allowed in db.txt

Change-Id: I5c8260872ffbc278d607f15899be53a96543aa93
CRs-Fixed: 651592