wlan: Process all candidate AP before setting state to not connected.
HDD connection state is set to eConnectionState_NotConnected for
each individual connection failure with the APs of the
requested profile i.e. with status eCSR_ROAM_ASSOCIATION_COMPLETION
and result eCSR_ROAM_RESULT_NOT_ASSOCIATED. Due to this if
disconnect command is sent from supplicant, it is dropped in HDD.
Eventually if DUT connects to an AP in the same requested profile,
driver is in connected state and supplicant is in
disconnected state.

Set HDD connection state to eConnectionState_NotConnected
only when CSR has completed all the available candidate AP
for the requested profile and Device is not able to
connect to any of them i.e status
eCSR_ROAM_ASSOCIATION FAILURE.

Process the Disconnect commands, even if SME is waiting
for the key.

Conflicts:
	CORE/HDD/src/wlan_hdd_assoc.c
	CORE/HDD/src/wlan_hdd_cfg80211.c
	CORE/SME/src/sme_common/sme_Api.c

Change-Id: I5edf92593fa765ace0ee43ceff207b64f33f69d6
CRs-Fixed: 668635