wlan: HDD_VOSS: Restrict creation of multi-port concurrency
Checks are added to restrict the creation of multi-port concurrency.
As of now, only two-port concurrency is supported; if there exists
a 2-port concurrency then the third connection is not allowed.
However, it does not restrict the creation of interface.
For example:
1. STA + P2P (GO/CLI) connection exists and then new
   connection establishment is not allowed.
2. STA + SAP connection exists and then new
   connection establishment is not allowed.
3. P2P GO + P2P CLI connection exists and then new
   connection establishment is not allowed.
4. STA + ADHOC connection exists and then new
   connection establishment is not allowed.
In the above example combinations, if any of the existing connection
is disconneted, then next connection is allowed.

CRs-fixed: 669604
Change-Id: Ib9e62953f38b274086115ae6a956ecd7a3973b0b