prima: Fix to avoid aging out connected BSS
We age out connected BSS as per current logic. Logic to age out
BSS is executed for each session. For the connected BSS in STA mode,
age out logic is not triggered for session 0. Since we check the
logic for each session, we proceeded further and age out logic got
triggered for session 1 which is not correct. This fix ensures to
avoid aging out connected BSS on any session and return from the
function once connected BSS is found.

Change-Id: I8865216ee9e5411096d79a774ad44108007a5c4f
CRs-Fixed: 689098