wlan: Avoid the race conditon when deauth ind is stuck.
Currently, when Deauth is received by AP and OS, deauth Ind
is stuck in CSR and deauth Req is not processed by lim and
LIM sends the deauth resp to SME.As a result, LIM clean up
does not happen and peer is not removed.When STA connects to
AP again, it crashed.
Changes are done to process the Deauth frame and remove the
peer to avoid the crash.

Change-Id: I4f145c4515ad4dde3d516d05a3af80eea3c806a7
CRs-Fixed: 689824