wlan: cap the TX power limits by the regulatory limits in NV
db.txt provides power limits per band per country, but a
mechanism should be provided so that the power limits are
capped by  the h/w, s/w specific regulatory conformace results.

To achieve this cap the TX power values to the values programmed
in NV for a regulatory domain corresponding to the country of
operation. The conformance test results have to be poulated in the
NV for all the supported regulatory domians as per this arrangement.

Change-Id: Iaf91486556dec9f98298caa60f0eeb438eac7080
CRs-Fixed: 689511