wlan: Fix to avoid SAP operating in DFS channel.
In ACS, SAP was never picking up channels based on the country code
acquired after the weights are computed for different channels. This
commit ensures SAP to pick the least congested channel based on the
country accquired.

Also, if all of the channels in the specified ACS range are disabled,
currently SAP will pick a default channel 6 if band selected is 2.4G
and default channel 40 if band is 5G. This is an issue in a scenario
where the specified ACS range is 5G and the country accquired has only
2.4G channels. As SAP would be picking an invalid channel 40 by default.
To address scenario, pick a default channel from 2.4G as 2.4G channels
are present in all the countries.

CRs-Fixed: 704735
Change-Id: I1b399d24f65af8799c1b17714a9f5b4e22ff8d0d