Accept RoC even if sta interface is in connecting state.
In concurrency cert test cases, RoC is rejected by the driver
because the driver is in connecting state for wlan0.
This is leading to failure as the peer device is not able
to discover the DUT.

Accepting the RoC even wlan0 is in connecting state. Driver dont
accept the RoC if p2p device is in connecting state.

Change-Id: Ie7410e26f6ab87624cb470c5551384f4a71579d2
CRs-Fixed: 708333