wlan: Flush Stale scan results for PNO Scan.
When scan is issued from host and simultaneously if PNO
indication is given to FW because of Apps power collapase;
A situation arises in FW where PNO can't be triggered because of
host scan. When host scan is over and there are many bssid
for same profile and connect is issued to such profile, it
might happen that because there are many stale entries connection
to bssid fails.But the PNO service manager in FW only waits for
10 sec for PNO to trigger,resulting in FW crash.

As a part of Fix, as part of PNO trigger flush stale scan result
in sme.

CRs-Fixed: 718160
Change-Id: I4303079eb49392859bdef59677c673629955bb9a