Wlan: If RSN IE is present ignore WPA IE to get cbmode
if AP has RSN IE and supports AES, driver goes on to check
WPA IE and if it contains only TKIP in unicast cipher, it
sets cbmode to 0 (20Mhz).

Added a check to ignore the WPA IE if RSN IE is present and take
decision on cbmode with respect to RSN IE only.

CRs-Fixed: 739345
Change-Id: Ifa973d9ed2599c60800c216c26fcc0cef860b130