wlan: Fix crash when open security & no traffic is running
In Open security case when there is no traffic is running,
it may possible tx time-out may once happen and later we
recovered then we need to reset the continuousTxTimeoutCount
because it is only getting modified when traffic is running.
So if over a period of time if this count reaches to
HDD_TX_STALL_SSR_THRESHOLD  then host is triggering false
subsystem restart.so in genuine Tx Time out case kernel will
call the tx time-out back to back at interval of HDD_TX_TIMEOUT.
So now we are checking if previous TX TIME out was occurred more
then twice of HDD_TX_TIMEOUT back then we may recovered here.

Change-Id: I284ccff253d89265d7874bc08d2db07bb24f93f5
CRs-Fixed: 737229