wlan: Abort Scan for Miracast session
Currently if split scan is going on and Miracast
session is started in between. Miracast session
doesn't allow scan on DFS channel and if before
Miracast starts as part of split scan if the
channel on which scan is intended is DFS , then
Channel number  0 is passed to lim. Since Channel
number is 0, lim issue full scan. As a part of fix,
abort any scan when Miracast Session is started.

Change-Id: I97d77772dbd2efdb4f07f84faa255686a2e8281f
CRs-Fixed: 740241