wlan:Set the hdd state to not connected even if disconnect timeout.
If disconnect command is queued in SME and same time the connect
is in progress which failed eventualy, driver will not call the
disconnect handler leading to the wait for disconnect handler
timeout. This timeout will leave the hdd state as disconneting
and thus the next connect commands will always try to disconnect
thinking that disconnect is still in progress leading to connect
return with failure.

With this fix hdd will set the hdd state from disconnecting to
not connected even if disconnect timeout.

Change-Id: Ib4eb403872b435746c598f162195194e32ea43f7
CRs-Fixed: 763699