wlan: moved ipv4 and ipv6 reg/unreg to load unload function
Unregistring of IP notifier was not happening properly
incase of unloading when in IBSS mode of operation.

To fix this in general not depending on mode of operation,
moved ipv4 and ipv6 registration and unregistration to load
unload function. With this, one single notifier registration
will take care of offloads depending on the adapter for which
the notifier callback is called.

Change-Id: Ica461f57c7eda09ca4bead9774379e2c5a59b0f6
CRs-Fixed: 774895.