wlan: Don't set CB mode in 2.4gz if FW doesn't have same capability.
Currently, for sta side if AP beacon contains HT ie and if its
secondary channel is valid, CB in host is set without checking
if FW supports CB in 2.4gz. As a part of fix make sure that prior
to setting CB mode in 2.4gz, FW capability is checked.

Change-Id: I54a634ad9b13865837727b248c7e66564e5d82a2
CRs-Fixed: 778209