wlan: avoid TX thread wait for haldumpcmd completion
In case of tx timeout we try to recover stall condition by
reloading FW. Haldumpcmd is sent to do this which is executed by
TX thread and response is expected to be processed by MCthread.

A race condition is observed when MC thread is unable to send
probe request on each channel due to TX thread stuck which is
due  to hdd_tx_timeout invoking haldumpcmd. TX wait for 10sec
until rsp for haldumpcmd is received.

Avoiding TX thread to wait for haldumpcmd RSP as not needed.

Change-Id: I1067c5213ac3424c6b3bab36c68cd953b4eb66bc
CRs-Fixed: 778948