wlan: wdi: prevent sending reset FIQ when ssr is in progress
SSR framework cannot handle new SSR request for a subsystem while
restart request for same subsystem is already in progress.
On WDI timeout we issue reset FIQ and wait for response.
On response failure we trigger SSR which cannot be handled by
SSR framework. Therefore we won't start timer for FIQ response
when SSR is already in progress.

Change-Id: I6f26be54509002cac5f0fac26f8c547eeacab9f3
CRs-Fixed: 799528