wlan:check for duplicate rates
Currently, if AP advertize the same rate in supported
and extended supported rate, driver takes that rate
in both the supported and extended supported rates.
This was crossing the max number of rates
which is 12 and leading to crash.To fix it include
check for any duplicate rates if exists then consider
once.

Change-Id: I073ee2ddf2b983865a329f51768bffdf4ca11c42
CRs-Fixed: 775179