wlan: Allow apps power collapse even when sco call is in progress
As a part of apps power collapse, Host check if BMPS is enabled
in ini then pmcState need to be BMPS else it rejects the request.
So when SCO indication comes from Coex module , host change
pmcState to full power mode hence it is not able to power collapse.

As part of this fix Host allows the power collapse now.

Change-Id: Ib6d9d77bd703ccc06de3f6d50f75d2896e39eacf
CRs-Fixed: 804864