wlan: Reinitialize the timer value during BTC indication
Currently, the time value is not re-initialized during
BTC indication. As a result, value of zero is passed on
to the subsequent module and the timer fails to start
since the timer value is zero. This is fixed as part of
this change.

Change-Id: I9010b24ac87a1ea90e28ca05d738cbce701395a6
CRs-Fixed: 808111
(cherry picked from commit fae117a3c69a76ce43b9dd29501e6db2c5dced6d)