wlan: Ignore the Ext Cap from supplicant if interworking is not set.
Driver is only interested in the interworkingService bit
of Ext Cap received from supplicant. So in case interworkingService
bit is not set in Ext Cap received from supplicant, it should not
be merged with the Ext Cap prepared by driver.

Added a check to ignore Ext Cap from supplicant if interworking bit
is not set.

CRs-Fixed: 810398
Change-Id: I0657589d49a632c044f4252d6cde705a9df2fef2