wlan: remove DXE Health Monitor
The DXE health monitor when triggered can write
into DXE registers leading to synchronization
issues with the hardware and software state.
The logs being output by the code are WARN
level logs which are never printed by default
and with all the enhancements in firmware and
the driver code are not necessary.

Change-Id: I851ef3b2b17488044df24921378e2394fc929bd9
CRs-Fixed: 830059