wlan: Optimize the code for validating the number of rates.
Currently there are two issues while populating the rates during
tdls add sta.

1. We simply return if number of B rates are greater the desired
value, instead of populating other rates.

2. Since we are post incrementing the A rate and B rate indices,
the B Rate index is starting from 1 to 4 which is validated as >=
to the desired value (4) and the condition is always met.

To address this issue, optimize the code for validating the
number of rates.

CRs-Fixed: 846356
Change-Id: I3684a496ea07be79fab9334fa163b0f93a10d734