wlan: increase the size of DXE debug array.
The DXE Stall debug array can hold information
for only 5 channels but DXE has 8 channels so
every time we add a new channel the array size
has to be increased to avoid memory corruption.
This change can be easily missed, so I have
increased the array size to 8 so that we don't
have to worry about memory corruption.

Change-Id: I8612dde2547ca5bb35884bbc38319b427b4330ba
CRs-Fixed: 848291