wlan: IBSS: Flush the scan list when IBSS is disconnected.
When IBSS is started, driver add the self info in the scan list.
If IBSS is stopped, the self entry is not removed from the scan
list and during scan the scan list shows the entry for the stopped
IBSS in the scan list for the gScanResultAgeCount times.

To avoid this flush the Scan list when the IBSS is stopped.

CRs-Fixed: 849842
Change-Id: I63084ec5cc449f3527cdf4a9b5f2bbdeed830e2c