prima: Ignore hdd_tx_timeout count when EAPOL is progress
When hdd_tx_timeout occurs, ignore continouous tx timeout count when
EAPOL is in progress. When EAPOL is in progress, the sta is in connected
state and TL will not fetch any frame till sta moves into authenticated
state. So timeout count has to be ignored else it will lead into SSR.

Change-Id: Ie8f74379c20584c6c4ab980dbcbe2c95d1b9fa12
Crs-Fixed: 872015