wlan: Fix Disconnection timeout.
Currently disconnection timeout is 2sec.
So if back to back connection and disconnection
is sent to driver, connection command might take
2sec to timeout in lim for assoc/auth resp. In
the meantime, disconnection command will timeout
and disocnnect will be sent to supplicant.This
results in supplicant and driver state mismatch.
As a fix, increase disconnection timeout. Also
if there is any disonnection from supplicant,
abort joining next bssid in csr.

Change-Id: I722aa2412f2b1b1d2343522f383f2f77c6c228c2
CRs-Fixed: 870857