Wlan: Drop connect after scan for ssid if discon is already issued
During connection if AP is not found in scan list, driver tries
to scan for the ssid (eCsrScanForSsid) and if AP is found,
the connect cmd is queued in sme queue.

During this scan if disconnect is issued from supplicant, as there is
no ongoing connect command and sme is not in connected state, the
disconnect command is dropped in sme. This leads to HDD connState
changed to eConnectionState_NotConnected. In this case driver tries
to abort the scan for ssid to stop the connect command.

But a condition may arise when scan for ssid is already completed
before abort was issued and thus connect command will be issued.

If this connect command is successful, it will move sme to
connected state but as hdd is in state eConnectionState_NotConnected
state the hdd assoc completion callback will drop this connect
indication from sme.

Now if supplicant sends a second connect command, with hdd in
eConnectionState_NotConnected state, it will be sent to sme.

If this second connect is for different BSSID, csr will continue
connect and ADD BSS will be sent with new AP's mac. This will lead
to FW crash as already a different bss entry is present in FW.

To avoid this don't issue connect after scan for ssid,
if disconnect is already initiated from Supplicant

Change-Id: Ie30d97f2da96baab03e3c072522500e97382671c
CRs-Fixed: 953404