Wlan: Decrements no. of active session for back to back connects
If device is in connected state and again connect is received from
supplicant, driver will disconnect the previous connection and
continue with connect. The no. of active session is not
decremented in this disconnect initiated from connect. Further
connect increment the no. of active session.

Now if one more connect is received driver assumes that max
concurrent connections has reached and return failure.

To avoid this, decrement no. of active session in case of
disconnect initiated from connect.

Change-Id: I5048fd76695890e3c84ba50315b1bcfdd4fcb91c
CRs-Fixed: 956934