Wlan: Avoid skipping DFS channels based on channel count
If only 5Ghz profiles are saved, as part of selective scan feature
only few 5ghz channels are received in the first scan. If skip dfs
channels in the fist scan is enabled and all the 5ghz channels
received in scan req are DFS/passive, number of channel sent to
lim is 0 and thus lim does the scan on all channels and Thus it
takes about 5-7 sec to display result on GUI.

Don't skip DFS channels in the first scan if the number of channels
in scan request is less than preconfigured value.

Change-Id: I50bd9920602c418ac4c0b7ac46301641c19a8636
CRs-Fixed: 964504