prima: Fix duplicate channel in RRM response
qcacld-2.0 to prima propagation

When non-connected AP moves from lower channel to higher channel, BSSID
information of that AP is sent in both old and new channels in RRM
response which is a spec violation. This is seen when RRM request is
made with active scan mode. Driver uses scan cache to send results as
part of RRM response. In scan cache, old entry is not removed by the
time RRM is requested after AP has moved into new channel and hence
duplicate info is seen in both channels. Fix this by comparing with
fresh scan results.

Change-Id: I7183ee7e8ec318078a1fc6d7d0c6c5362d5cf498
CRs-Fixed: 676619