wlan: Drop Actions frames in FW, that are not processed by host
Several Action frame categories are not processed in host and they
are not forwarded to supplicant either, but they are dropped in host.
These unprocessed action frames are forwarded to host by FW and
wake up host unnecessarily which leads to battery drain. With this fix,
host conveys FW the allowed/required action frames list at HDD
startup so that the remaining frames can be dropped in FW itself
thereafter.

Change-Id: I97b87c9fa04973a5edbd0a4aa3aa5dd7c3d3e039
CRs-Fixed: 999114