wlan: Fix for sending action frame in SAP mode
In SAP mode when action frame is sent on the same
operating channel in which SAP is started and if
hostap sends a non zero wait time , driver is not
able to send the action frame.

To address this, in SAP mode, set the wait as zero,
so that the action frame is sent on the same
channel on which the sap is started.

CRs-Fixed: 1030645
Change-Id: I3a127fd704815ffb28a51bd9961287a7c62a815c