Prima: Add Flag to mark Dis-association in progress for STA
Currently, TL state for STA moves the suspend, when dis-association
is received from AP. Once dis-association indication reach to HDD,
it clear STA contex. In between packets for that STA reaches to TL
and found TL suspended with valid STA context. Dut to this packet
stuck in TL queues and Watchdog barks. To avoid that added a flag
to mark dis-association in progress for perticular STA. So TL will
avoid this STA to send packet.

CRs-Fixed: 1040417
Change-Id: I1d06232f986109984e3cc897f26ba5d3a6f36c1a