wlan: Add support to disable forwarding of BAR frames
In suspend mode, if AP sends the BAR frames then these BAR
frames are sent to host which leads to unnecessary
APPS wakeup.
Add ini support to disable the BAR frames. When this ini is set
to 1, FW will not forward the BAR frames to host when APPS is
in suspend mode.
Change-Id: Ifc4e4f3ea0c20f7073f4ee5b01b0a7681225e4b4
CRs-Fixed: 948786