wlan: HAL changes for dropping Actions frames in FW
Several Action frame categories are not processed in host and they
are not forwarded to supplicant either, but they are dropped in host.
These unprocessed action frames are forwarded to host by FW and
wake up host unnecessarily which leads to battery drain. With this fix,
adding HAL interface changes for host to convey FW the allowed/required
action frames list at HDD startup so that the remaining frames can be
dropped in FW itself thereafter.

Change-Id: Ib267bdb82b172a2de72fe166a318f218e672713d
CRs-Fixed: 999114