wlan: Add check to not trigger LFR MBB if disconnected
If deauth is received before lfr mbb is kicked in and while
deauth processing in progress, if lfr mbb is kicked in, then
there is a possibility of race. Add a check to avoid triggering
LFR MBB if CSR state is eCSR_NEIGHBOR_ROAM_STATE_INIT. Also,
set is_preauth_lfr_mbb as soon as handoff is triggered and
drop candidate indication if deauth/disassoc in progress.

Change-Id: I384a66ae50a6d143cfe5c985de33610af8f920fd
CRs-Fixed: 2006380