wlan: Set proper scan status in csrScanSmeScanResponse
On receiving sme scan response for successful ssid scan,
status is not set to success which results in
csr_ssid_scan_done_callback to send association failure
indication to HDD and set connection state to not_connected.
Since connection state is set to not_connected, connection
indication is not sent to userspace even though connection
is successful. Supplicant will issue connect again where
connection state is set to connecting. If the connect req
is for same profile, driver sends connection indication to
userspace but if it is for different profile connect req
fails. This will result in subsequent scans getting rejected
since connection state is stuck in connecting.

Fix is to set the proper scan status so that association
indication is sent properly to HDD and userspace.

Change-Id: I1e7264e71f1878bbba3d6d7ffda4a3820c94ac62
CRs-Fixed: 2177974