wlan: Match all encryption type in case of testbed STA
If force_rsne_override is set i.e testbed STA mode, the auth type
and encryption type set in RSNE provided, may not match with the
AP. In this case driver will not find the AP in filtered scan
result using the encryption and auth type provided in RSNE.
Thus no Assoc req is sent to AP and the test case fails.

As part of the fix set all authtype and RSN encryptions in filter
so that driver is able to find the AP and send Assoc request.

Change-Id: If99df221b10c5276f902df2f0a52b87e4df1be2c
CRs-Fixed: 2263295